refactor: move contracts functions to rpc
